feat(packages/sui-pde): add getDecision function
