+++ packages/sui-pde/src/index.js
@@ -2,6 +2,7 @@ export {default as PDE} from './pde.js'
 export {default as useFeature} from './hooks/useFeature.js'
 export {default as PdeContext} from './contexts/PdeContext.js'
 export {default as useExperiment} from './hooks/useExperiment.js'
+export {default as getDecision} from './hooks/common/getDecision.js'
 export {default as useDecision} from './hooks/useDecision.js'
 export {default as useDecisionCallback} from './hooks/useDecisionCallback.js'
 export {default as Experiment} from './components/experiment.js'
